Move Sorting Functions to Separate Module
